add files via upload
Waldo was an idea where the software can find waldo in a given image. The problem is, the quality on the images online are terrible and I am not sure how well the program will perform on small images. Also, I would need to teach it what waldo looks like because he does not look like a human face